refactor(manage): refactor crown development include logic and add new payload contracts
